My_Project/CRM/pyproject.toml
[project]
name = "crm"
version = "0.1.0"
description = ""
authors = [
    {name = "Sp1ker2",email = "bogdan.speaker@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.16.5,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
